finder: overhaul resource collection and aggregation
Simplify the way we gather container resource and
overhaul the aggregation phase.

Temporarily disable the device availability reporting:
there is no generic solution in sight, for example there
is no guarantee a device ID is a PCI address, or even
that allow us to find the PCI address of a device, and
the NUMA locality (NUMA NODE id) of a device is not exposed.

So we need to add device-plugin specific code, which will
be added in a future patch.

Signed-off-by: Francesco Romani <fromani@redhat.com>