finder: report SRIOV devices availability
reporting device-related availability is an hard task at current time.

Let's consider a generic device plugin which manages "resourceABC"
arbitrating access to a PCI device (generalizing GPU and SRIOV cases)

Challenges are:
- there is no easy and supported way to enumerate resources, both
  per-numa and in total. While the device plugin and the kubelet have
  this information, it is not exposed outside the kubelet.
- there is no easy and supported way to do the reverse mapping between
  resource name and PCI address. We use the PCI address because this is
  safe proxy for the NUMA node locality of the resource being
  considered. Even better would be to have a way to learn the NUMA
  locality of the resources (perhaps per device ID), but again while
  device plugin and kubelet both have this information, it is not
  exported.

So, we can easily learn that pod p1 and its container c1 are given
exclusive access to device d1, but there is no easy and supported way to
learn to which NUMA node the device d1 is affine to.
This in stark contrast with the CPU affinity, which is pretty easy (and
safe) to know.

At the best of our knowledge, to actually fix these issues we need API
changes in the kubelet, perhaps in the CRI API, perhaps in the
podresources API, perhaps in both.

To move forward in the near term, this patch implements a series of
changes intended to work with the SRIOV device plugin only, which is at
this time the device class we care most about.

First and foremost, we require the user to give to RTE a mapping between
the resource name and the *parent* SRIOV device (aka the physical
function - PF). If needed, it is easy to let RTE fetch the data from the
SRIOVNodeState object as future extension.

Once we know the mapping between resource name and the SRIOV PF, we can
track all the Virtual Functions (VFs) which are created in the system
from these PF. We can now
1. sum all the VFs, thus have a very good estimate of the resource
   capacity
2. learn about the NUMA locality of each VF, using the VF virtual device PCI address
   (handled by the linux kernel)

The final missing link is to learn which VF is assigned to which
container. To learn that we leverage the fact we know how the SRIOV
device plugin let the container know about the device.
It creates a well known environment variable inside the container, whose
value is the PCI address(es) of the device(s) assigned to the container.

The RTE now scans for these variables to learn about the devices
assigned to containers; from the PCI address we can find the VF (thus
the PF, hence the resource name) and the NUMA locality, so we can do
proper accounting.

Signed-off-by: Francesco Romani <fromani@redhat.com>